Move custom waiter tests into main test file
